MusicOnHold Functionality Added
Added music on hold functionality when there is one channel alone in the conference. When a channel joins the mixing bridge they are played MoH, but it stops when another channel joins the bridge. Also if there are multiple people in the bridge and everyone, besides one person, decides to leave then that one person left in the bridge will hear MoH.