Split up code to remove clutter.
Added a data section to the repository. This will contain anything and
everything that has to do with data access, and all that needs to be
called is the surface file (db.js) which contains all the functions
needed by the app to access the database. On top of that, a helper file
was added to do all the heavy lifting for the app behind the scenes.

* Added data access layer
* Added a helper file to place users into the bridge
* Significantly reduced the workload of confbridge.js