Added PIN authorization and bridge profiles.
After setting up a basic conference, we added PIN authorization to
help restrict which users are able to join the conference. While
PIN authorization is specified under user profiles, the PIN number
and retries are located in the bridge profiles, which we also added
to our application. The bridge profile for the application can be
specified in the config.json file.

* Added PIN authorization
* Added bridge profiles