Added the DTMF menu functionality
This update includes all of the DTMF
functionality needed within a conferencing
application, minus that of recording.
This functionality includes the likes of
decreasing, increasing, and resetting the
listening and talking volumes. As well as
being able to mute, deaf mute, change the
pitch of ones voice, and exiting the conference
in order to execute something within the dialplan.

* Added the menu functionality