Added minor group functionality.
We added the ability to provide a group profile to the Stasis
command, which will in turn fetch the profile from the database.
A user will default to a participant if nothing is specified,
which enters the application like normal. Leaders and followers
are kept up with in groupsetup.js. Followers are put into the
waiting state (waiting for leaders). Leaders have no functionality
as of right now.

* Added database retrieval for groups
* Added a module to keep track of groups
* Added minor follower functionality