Finished basic group functionality.
Added leader and follower functionality. When a follower joins the
conference, they will be placed into a waiting state until a leader
joins, unless there is a leader already present. Leaders enter the
conference like normal and serve primarily as a trigger for
followers. Groups are configured in the database, just like users.
We also cleaned up some code here and there.

* Added leaders/followers/participants
* Did some minor code clean up