Added data scripts, documentation, and cleanup.
We added another directory containing scripts to run to set up
the database. One creates or resets the database to default
tables, and the other three add a row to a specific table in
the database.
There's now a documentation section! We added content to the
README file, explaining how things work and what's available
in the database.
We also cleaned up some content within the application to help
with readability.

* Added database scripts
* Added documentation
* Cleaned up code