Fixed some minor issues/typos.
There were a couple issues with followers that needed to be
addressed before merging. A join / leave sound would be heard by
followers at inappropriate times because of how they enter and
leave the conference (via leader presence). There was also a typo
in the defaultdatabase script that would cause followers to not
hear the 'waiting for leader' prompt (oops).

* Fixed join/leave sounds for followers
* Fixed a typo in database scripts